update: fix bug and  access writes

+++ monsupersite/App/Backend/Modules/News/NewsController.php
@@ -11,25 +11,35 @@ use \OCFram\FormHandler;
  
 class NewsController extends BackController
 {
-  public function executeDelete(HTTPRequest $request)
-  {
-    $newsId = $request->getData('id');
- 
-    $this->managers->getManagerOf('News')->delete($newsId);
-    $this->managers->getManagerOf('Comments')->deleteFromNews($newsId);
- 
-    $this->app->user()->setFlash('La news a bien été supprimée !');
- 
-    $this->app->httpResponse()->redirect('.');
+  public function executeDelete(HTTPRequest $request){
+
+    if($this->app->user()->getUser()->fucType() == 1){
+        $newsId = $request->getData('id');
+     
+        $this->managers->getManagerOf('News')->delete($newsId);
+        $this->managers->getManagerOf('Comments')->deleteFromNews($newsId);
+     
+        $this->app->user()->setFlash('La news a bien été supprimée !');
+     
+        $this->app->httpResponse()->redirect('.');
+    }
+    else{
+        $this->app->httpResponse()->redirect('.');
+    }
   }
  
   public function executeDeleteComment(HTTPRequest $request)
   {
-    $this->managers->getManagerOf('Comments')->delete($request->getData('id'));
- 
-    $this->app->user()->setFlash('Le commentaire a bien été supprimé !');
- 
-    $this->app->httpResponse()->redirect('.');
+    if($this->app->user()->getUser()->fucType() == 1){
+        $this->managers->getManagerOf('Comments')->delete($request->getData('id'));
+     
+        $this->app->user()->setFlash('Le commentaire a bien été supprimé !');
+     
+        $this->app->httpResponse()->redirect('.');
+    }
+    else{
+        $this->app->httpResponse()->redirect('.');
+    }
   }
  
   public function executeIndex(HTTPRequest $request)
@@ -58,36 +68,53 @@ class NewsController extends BackController
  
   public function executeUpdateComment(HTTPRequest $request)
   {
-    $this->page->addVar('title', 'Modification d\'un commentaire');
- 
-    if ($request->method() == 'POST')
-    {
-      $comment = new Comment([
-        'id' => $request->getData('id'),
-        'auteur' => $request->postData('auteur'),
-        'contenu' => $request->postData('contenu')
-      ]);
+
+    if($this->app->user()->getUser()->fucType() == 1){
+        $this->page->addVar('title', 'Modification d\'un commentaire');
+     
+        if ($request->method() == 'POST')
+        {
+          if($request->postData('averti')){
+              $bool = 1;
+          }
+          else{
+              $bool = 0;
+          }
+
+          $comment = new Comment([
+            'id' => $request->getData('id'),
+            'news' => $request->getData('news'),
+            'mail' => $request->postData('mail'),
+            'auteur' => $request->postData('auteur'),
+            'contenu' => $request->postData('contenu'),
+            'averti' => $bool,
+          ]);
+        }
+        else
+        {
+          $comment = $this->managers->getManagerOf('Comments')->get($request->getData('id'));
+
+        }
+     
+        $formBuilder = new CommentFormBuilder($comment);
+        $formBuilder->build();
+     
+        $form = $formBuilder->form();
+     
+        $formHandler = new FormHandler($form, $this->managers->getManagerOf('Comments'), $request);
+     
+        if ($formHandler->process())
+        {
+          $this->app->user()->setFlash('Le commentaire a bien été modifié');
+     
+          $this->app->httpResponse()->redirect('/admin/');
+        }
+     
+        $this->page->addVar('form', $form->createView());
     }
-    else
-    {
-      $comment = $this->managers->getManagerOf('Comments')->get($request->getData('id'));
+    else{
+        $this->app->httpResponse()->redirect('.');
     }
- 
-    $formBuilder = new CommentFormBuilder($comment);
-    $formBuilder->build();
- 
-    $form = $formBuilder->form();
- 
-    $formHandler = new FormHandler($form, $this->managers->getManagerOf('Comments'), $request);
- 
-    if ($formHandler->process())
-    {
-      $this->app->user()->setFlash('Le commentaire a bien été modifié');
- 
-      $this->app->httpResponse()->redirect('/admin/');
-    }
- 
-    $this->page->addVar('form', $form->createView());
   }
  
   public function processForm(HTTPRequest $request)
