+++ monsupersite/App/Backend/Modules/User/UserController.php
@@ -13,19 +13,31 @@ class UserController extends BackController
 
 
    public function executeIndex(HTTPRequest $request){
-    $this->page->addVar('title', 'Gestion des Ecrivain');
- 
-    $manager = $this->managers->getManagerOf('Users');
- 
-    $this->page->addVar('listeEcrivain', $manager->getListeEcrivain());
+    if($this->app->user()->getUser()->fucType() == 1){
+        $this->page->addVar('title', 'Gestion des Ecrivain');
+     
+        $manager = $this->managers->getManagerOf('Users');
+     
+        $this->page->addVar('listeEcrivain', $manager->getListeEcrivain());
+    }
+    else{
+        $this->app->httpResponse()->redirect('../');
+    }
   }
 
   public function executeDelete(HttpRequest $request){
-  	$manager = $this->managers->getManagerOf('Users');
 
-	$userId = $request->getData('id');
-  	$manager->delete($userId);
-  	$this->app->httpResponse()->redirect('.');
+    if($this->app->user()->getUser()->fucType() == 1){
+      	$manager = $this->managers->getManagerOf('Users');
+
+    	  $userId = $request->getData('id');
+      	$manager->delete($userId);
+      	$this->app->httpResponse()->redirect('.');
+    }
+    else{
+        $this->app->httpResponse()->redirect('../');
+    }
+
   }
